requirements.txt
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
cryptography==46.0.7
Flask==3.1.3
gunicorn==25.3.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.2
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
